Classes table auto-generated from fake data
Generated fake data in views.py and passed into classes.html template.
Used fake data to generate class table. Added javascript functionality
to the buttons, with alerts verifying that the proper data is being
sent through the buttons. Added a stylesheet 'style.css' and linked it
to the master templates. Added an h1 tag in the master-instructor
template for page titles.