Finished front-end basics for all pages
All pages are now formatted with CSS and have Javascript event
handlers. Added an instructor signup page. Attached links to the
buttons in the navbar, added a login button, and made the logout button
actually log the user out. Added the CSS and javascript for the
Submissions page. Made several minor improvements in the formatting of
each page. Adjusted some of the routing for home pages in views.py.