Database can now add classes and assignments
Linked the create assignment and create class pages to the database.
Added modals to report on the success/failure of the create-class and
create-assignment functions. Added a consistent format string for
displaying due dates. Modified the safe_division function to return 0
when dividing by 0 (so that if there are no students in a particular
class, the current-assignments page will display 0% submitted and
refrain from creating a "View" button.