Submissions can now be added to the database
All operations that add items to the database are completed.  Added a
modal to view a list of students for each class on the classes page.
Added a modal to view each submission on the submissions page. Fixed
the percentage calculation (for determining percentage of class that
has submitted an assignment) by forcing floating point division. Made
several CSS formatting adjustments. Cleared all of the tables in app.db
and started fresh.