Current and past-assignments are now different
Added a filter to only display assignments with due dates today or
later as current-assignments, all else are displayed on the
past-assignments page. Upon creation of a new assignment or a new
class, instructors are redirected to the home page
(current-assignments). Also, added several assignments and submissions
to demonstrate the color-coding on the current and past-assignment
pages.